requirements_file.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
spacy==3.8.0
en-core-web-sm==3.8.0
requests==2.31.0
Werkzeug==2.3.7
mistralai==1.8.2
PyMuPDF==1.23.8
python-dotenv
passlib[bcrypt]
python-multipart
pymongo
